backend/tracking/middleware.py
"""
Custom middleware for JWT authentication in WebSocket connections.
"""
from channels.middleware import BaseMiddleware
from channels.db import database_sync_to_async
from django.contrib.auth import get_user_model
from rest_framework_simplejwt.tokens import UntypedToken
from rest_framework_simplejwt.exceptions import InvalidToken, TokenError
from jwt import decode as jwt_decode
from django.conf import settings
from urllib.parse import parse_qs

User = get_user_model()


class JWTAuthMiddleware(BaseMiddleware):
    """JWT authentication middleware for WebSocket connections."""
    
    async def __call__(self, scope, receive, send):
        """Process WebSocket connection and add user to scope."""
        # Extract token from query string or headers
        token = None
        
        # Try query string
        query_string = scope.get('query_string', b'').decode()
        if query_string:
            query_params = parse_qs(query_string)
            token = query_params.get('token', [None])[0]
        
        # Try headers if not in query string
        if not token:
            headers = dict(scope.get('headers', []))
            auth_header = headers.get(b'authorization', b'').decode()
            if auth_header.startswith('Bearer '):
                token = auth_header.split('Bearer ')[1]
        
        # Authenticate token
        user = None
        if token:
            try:
                UntypedToken(token)
                decoded_data = jwt_decode(token, settings.SECRET_KEY, algorithms=["HS256"])
                user_id = decoded_data.get('user_id')
                if user_id:
                    user = await self.get_user(user_id)
            except (InvalidToken, TokenError):
                pass
        
        # Add user to scope
        scope['user'] = user
        
        return await super().__call__(scope, receive, send)
    
    @database_sync_to_async
    def get_user(self, user_id):
        """Get user asynchronously."""
        try:
            return User.objects.get(id=user_id)
        except User.DoesNotExist:
            return None

