backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
python-dotenv==1.0.0
djangorestframework-simplejwt[blacklist]==5.3.0
django-filter==23.5
daphne==4.0.0
channels-redis==4.1.0

